% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.TVGEV}
\alias{anova.TVGEV}
\title{Analysis of Deviance Table for two Nested \code{TVGEV}
    Objects}
\usage{
\method{anova}{TVGEV}(object, object1, trace = 1L, ...)
}
\arguments{
\item{object}{A \code{TVGEV} object as fitted by
\code{\link{TVGEV}}.}

\item{object1}{A \code{TVGEV} object such that \code{object} is
nested in \code{object1}.}

\item{trace}{Level of verbosity. The value \code{0} prints
nothing.}

\item{...}{Not used yet.}
}
\value{
An object of class \code{"anova"} inheriting from class
    \code{"data.frame"}.
}
\description{
Compute an analysis of deviance table for two nested \code{TVGEV}
objects.
}
\note{
The deviance of the models can not be interpreted: only the
    difference of the deviance is used.
}
\section{Caution}{
 The distribution of the test statistic
    (difference between two deviances) is obtained on the basis of
    the \emph{large sample theory} which may not be applicable.
}

\examples{

df <- within(TXMax_Dijon, Date <- as.Date(sprintf("\%4d-01-01", Year)))

## fit without a break in trend
res0 <- TVGEV(data = df, response = "TXMax", date = "Date",
               design = breaksX(date = Date, breaks = "1970-01-01", degree = 1),
               loc = ~ t1,
               estim = "nloptr")

## the same but with a break
res1 <- TVGEV(data = df, response = "TXMax", date = "Date",
              design = breaksX(date = Date, breaks = "1970-01-01", degree = 1),
              loc = ~ t1 + t1_1970,
              estim = "nloptr")
anova(res0, res1)

}
