% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bts.R
\name{as.bts.data.frame}
\alias{as.bts.data.frame}
\title{Coerce a Data Frame into a \code{bts} object}
\usage{
\method{as.bts}{data.frame}(object, dateName, yearName, ...)
}
\arguments{
\item{object}{A data frame with one column}

\item{dateName}{Character. Optional column name for a date in an
understandable format.}

\item{yearName}{Character. Optional column name for a year. When
given, the column must be either numeric or character.}

\item{...}{Note used.}
}
\value{
An object with class \code{"bts"} inheriting from
\code{"matrix"}.
}
\description{
Coerce a data frame into a \code{bts} object, using a suitable
column as a date.
}
\examples{
## does not work
bts0 <- try(as.bts(TXMax_Dijon))
## works
bts1 <- try(as.bts(TXMax_Dijon, year = "Year"))

}
