% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bts.R
\name{as.bts}
\alias{as.bts}
\title{Coercion into a \code{bts} Object}
\usage{
as.bts(object, ...)
}
\arguments{
\item{object}{An object to be coerced into a \code{bts} object.}

\item{...}{Extra arguments for methods.}
}
\value{
An object with class \code{"bts"}.
}
\description{
Coercion into a \code{bts} object.
}
