% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bts.R
\name{as.data.frame.bts}
\alias{as.data.frame.bts}
\title{Coerce an \code{bts} Object into a Data Frame}
\usage{
\method{as.data.frame}{bts}(x, row.names = NULL, optional = FALSE,
  longFormat = FALSE, ...)
}
\arguments{
\item{x}{An object with class \code{bts} representing a by-block time
series.}

\item{row.names}{Not used.}

\item{optional}{Not used.}

\item{longFormat}{Logical. With the default \code{FALSE}, the
result has the same columns as \code{x} plus the new column
\code{Date}. When \code{longFormat} is \code{TRUE}, the data frame
is in "long format" with all the columns of \code{x} given in same
the column \code{value} of the result. A column \code{type} indicates
the name of the original column in \code{x}.}

\item{...}{Not used yet.}
}
\value{
A dataframe with a \code{Date} column containing the value
of the \code{"date"} attribute of \code{x}.
}
\description{
Coerce an \code{bts} object into a data frame
}
\examples{
myDate <- seq(from = as.Date("1996-01-01"),
            to = as.Date("2016-12-31"), by = "years")
X <- polynomX(myDate)
df <- as.data.frame(X)
head(df)
}
