% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asNSGEV.R
\name{as.NSGEV.fevd}
\alias{as.NSGEV.fevd}
\title{Coerce a 'gev.fit' object into a 'NSGEV' object}
\usage{
\method{as.NSGEV}{fevd}(x, ...)
}
\arguments{
\item{x}{The object to be coerced. Must have clas \code{"fevd"} and
be of type \code{"GEV"}}

\item{...}{Not used yet.}
}
\value{
An object with class \code{"NSGEV"}.
}
\description{
Coerce a 'gev.fit' object into a 'NSGEV' object.
}
\section{Caution}{
 For now, only simple formulas can be used.
}
\examples{
require(extRemes)

## see the examples for extRemes::fevd.
data(PORTw)

## fit a GEV model
fit <- fevd(x = TMX1, data = PORTw,
            location.fun = ~AOindex, scale.fun = ~AOindex,
            units = "deg C")

## coerce
ns <- as.NSGEV(fit)
}
\author{
Yves Deville
}

