% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asNSGEV.R
\name{as.NSGEV}
\alias{as.NSGEV}
\title{Coerce into a \code{NSGEV} object}
\usage{
as.NSGEV(x, ...)
}
\arguments{
\item{x}{Object to be coerced.}

\item{...}{Special arguments that may be required for some
classes.}
}
\value{
An object with (S3) class \code{"NSGEV"} representing
a Non-Stationary GEV model.
}
\description{
Coerce into a \code{NSGEV} object.
}

