% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bts.R
\name{as.ts.bts}
\alias{as.ts.bts}
\title{Coerce a \code{bts} Object into a \code{ts} Time Series Object}
\usage{
\method{as.ts}{bts}(x, ...)
}
\arguments{
\item{x}{A \code{bts} object.}

\item{...}{Not used.}
}
\value{
An object inheriting from the \code{"ts"} S3 class.
}
\description{
Coerce a \code{bts} object into a \code{ts} time series object.
}
