% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotStuff.R
\name{autolayer.quantMax.TVGEV}
\alias{autolayer.quantMax.TVGEV}
\title{Draw a Particular Layer on a \code{ggplot} devoted to a
    \code{quantMax.TVGEV} object.}
\usage{
\method{autolayer}{quantMax.TVGEV}(object, which = c("quantile", "confBand"), ...)
}
\arguments{
\item{object}{A \code{quantMax.TVGEV} object.}

\item{which}{Character vector describing the}

\item{...}{Further arguments for the involved geometry.}
}
\value{
A layer to be added to a \code{ggplot} object.
}
\description{
Draw a Particular Layer on a \code{ggplot} devoted to a
    \code{quantMax.TVGEV} object. This method may be used to
    compare different fitted object or different periods of time.
}
\details{
When \code{which = "quantile"} is used the quantile line is added
   via a \code{geom_line} layer.  The arguments passed via the
   dots can be \code{colour}, \code{linetype}, \code{alpha},
   ... When \code{which = "confband"} the confidence band (or
   area) is added via a \code{geom_ribbon} layer. The arguments
   passed via the dots can be \code{fill}, \code{alpha}, ...
}
\seealso{
The \code{\link{autoplot.quantMax.TVGEV}} method. The
    \code{\link[ggplot2]{geom_line}},
    \code{\link[ggplot2]{geom_ribbon}} geometries.
}
