% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{autoplot.bfts}
\alias{autoplot.bfts}
\title{Autoplot for Block Functional Time Series \code{bfts} Objects}
\usage{
\method{autoplot}{bfts}(object, col = "orangered", fill = NA, ...)
}
\arguments{
\item{object}{A \code{bfts} object such as returned by the
\code{density} and \code{cdf} methods for \code{TVGEV} objects.}

\item{col}{Colour for the line and fill.}

\item{fill}{If non \code{NA}, the region between the curve and the
axis will be filled as is quite usual with probability densities.}

\item{...}{Not used yet.}
}
\value{
A \code{ggplot} object or nothing.
}
\description{
Autoplot for Block Functional Time Series \code{bfts} objects.
}
\note{
This function is mainly intended to be used for density
functions or cdf. In the former case, \code{fill = TRUE} may be
preferred.
}
