% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{autoplot.bts}
\alias{autoplot.bts}
\alias{autolayer.bts}
\title{Plot a Block Time-Series using ggplot}
\usage{
\method{autoplot}{bts}(object, col1 = "darkgray", facets = FALSE, ...)

\method{autolayer}{bts}(object, col1 = "darkgray", ...)
}
\arguments{
\item{object}{An object with class \code{"bts"}.}

\item{col1}{A colour to be used when the number of component
series in \code{object} is > 12.}

\item{facets}{Logical. If true, each component series will be
shown in a facet.}

\item{...}{Not used}
}
\value{
An object of class \code{"ggplot"}.
}
\description{
Plot a block time-series using ggplot.
}
