% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influenceTVGEV.R
\name{autoplot.influence.TVGEV}
\alias{autoplot.influence.TVGEV}
\title{Create a \code{ggplot} for a \code{influence.TVGEV} Object}
\usage{
\method{autoplot}{influence.TVGEV}(object, ...)
}
\arguments{
\item{object}{An object with class \code{"influence.TVGEV"} as
created bye the \code{influnce} method for the \code{"TVGEV"}
class.}

\item{...}{Not used yet.}
}
\value{
An object inheriting grom \code{"ggplot"} showing the
    (finite-sample) influence function for the observation and the
    statistic that have been chosen.
}
\description{
The influence curve is obtained by replacing the chosen
observation by a candidate value and then re-fitting the same
model and computing the statistic of interest: coefficient
estimate, quantile, ... The rug shows the \eqn{y_b} and the
vertical line shows the observation chosen for the analysis.
}
\examples{
library(ismev)
data(venice)
df <- data.frame(Date = as.Date(paste0(venice[ , "Year"], "-01-01")),
                 Sealevel = venice[ , "r1"] / 100)
fit0 <- TVGEV(data = df, date = "Date", response = "Sealevel")
autoplot(fit0)
RL_2050 <- function(model) {
    c("RL_2050(100)" = quantile(model, prob = 0.99, date = "2050-01-01")[1])
}
autoplot(fit0)
influence(fit0, what = RL_2050) |> autoplot()
## fit with a linear time trend
fit1 <- TVGEV(data = df, date = "Date", response = "Sealevel",
              design = polynomX(Date, degree = 1), loc = ~t1)
autoplot(fit1)
summary(fit1)
influence(fit1) |> autoplot()
influence(fit1, what = RL_2050) |> autoplot()
influence(fit1, which = "max", what = RL_2050) |> autoplot()
## Influence curve for the estimated slope
slope <- function(model) {
    c("slope\n(cm/100 yr)" = 100 * unname(coef(model)["mu_t1"]))
}
influence(fit1, which = "max", what = slope) |> autoplot()
influence(fit1, which = "min", what = slope) |> autoplot()
}
