% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotStuff.R
\name{autoplot.quantMax.TVGEV}
\alias{autoplot.quantMax.TVGEV}
\title{Create a \code{ggplot} for a \code{quantMax.TVGEV} Object}
\usage{
\method{autoplot}{quantMax.TVGEV}(object, fillConf = TRUE, ...)
}
\arguments{
\item{object}{An object with class \code{"quantMax.TVGEV"} as
produced by the \code{\link{quantMax}} method.}

\item{fillConf}{Logical. If \code{TRUE} the "confidence band(s)"
will be filled.}

\item{...}{Not used.}
}
\description{
Create a \code{ggplot} for a \code{quantMax.TVGEV} as created by
using the \code{quantMax} method.
}
\seealso{
The \code{\link{quantMax.TVGEV}} method.
}
