% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residTVGEV.R
\name{autoplot.resid.TVGEV}
\alias{autoplot.resid.TVGEV}
\title{Plot Residuals of a \code{TGVEV} Model}
\usage{
\method{autoplot}{resid.TVGEV}(object, geom = c("line", "segment"), ...)
}
\arguments{
\item{object}{A \code{TVGEV} object.}

\item{geom}{The "geometry" used to show the series of block
maxima. The default \code{"line"} correspond to a polyline and
\code{"segment"} correspond to a series of vertical segments
a.k.a. "needles".}

\item{...}{Not used yet.}
}
\value{
Nothing.
}
\description{
Plot the residuals of a \code{TVGEV} model object against the
date.
}
\seealso{
\code{\link{residuals.TVGEV}}.
}
