% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTVGEV.R
\name{autoplot.TVGEV}
\alias{autoplot.TVGEV}
\title{Create a \code{ggplot} for a \code{TVGEV} Object}
\usage{
\method{autoplot}{TVGEV}(object, geom = c("line", "segment"), type = c("quant"), ...)
}
\arguments{
\item{object}{An object with class \code{"TVGEV"}.}

\item{geom}{Character giving the "geometry" to be used for the
layer corresponding to the (timeseries) data.}

\item{type}{The type of result shown along with the series used.
For now, only the timeseries of quantiles as computed by
\code{\link{quantile.TVGEV}} can be added.}

\item{...}{Not used yet}
}
\value{
A graphic object inheriting from \code{"ggplot"}.
}
\description{
Create a \code{ggplot} for a \code{TVGEV} Object
}
\note{
Mind that several methods of the class \code{"TVGEV"} such
    as \code{predict} \code{quantMax}, \code{coef} produce results
    that can be autoplotted.
}
