% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{blockDuration}
\alias{blockDuration}
\title{Find Block Duration}
\usage{
blockDuration(date)
}
\arguments{
\item{date}{A vector with class \code{"Date"} or that can be
coerced to this class.}
}
\value{
A duration in years.
}
\description{
Find block duration in years.
}
\details{
This is simply a wrapper for the \code{diff} method.
}
\note{
For most applications of Time Varying GEV models with
\code{TVGEV}, the block duration will be one year. Yet it can be
sometimes needed to form blocks with longer duration (e.g. two
years, five years).
}
\examples{

date <- as.Date(sprintf("\%4d-01-01", TXMax_Dijon$Year))
n <- length(date)
blockDuration(date)
blockDuration(date[seq(from = 1, to = n, by = 2)])
}
