% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{breaksX}
\alias{breaksX}
\title{Design Matrix with Breaks in a Linear Trend}
\usage{
breaksX(date, degree = 1L, origin = NULL, breaks = NULL,
  constant = FALSE)
}
\arguments{
\item{date}{A vector with class \code{"Date"} or a vector that can
be coerced to the \code{"Date"} class, typically an unambiguous
character vector.}

\item{degree}{The degree used for spline functions.}

\item{origin}{Not used yet.}

\item{breaks}{A vector describing the breaks. It will be coerced
to the \code{"Date"} class, and can be e.g. an unambigous
character vector.}

\item{constant}{Logical. If \code{TRUE} a column with constant
unit value is added.}
}
\value{
An object of class \code{"bts"} inheriting from
\code{"matrix"}. This is essentially a numeric matrix with the
value of the basis functions as its columns. The rows are in
correspondence with the elements of the \code{date} vector given
on input.
}
\description{
Build a design matrix to describe breaks in a linear trend.
}
\details{
The design matrix is a Truncated Power Basis of splines. It can be
extended to a degree differing from the default value \eqn{1}.
The provided date is converted into a numeric variable \eqn{t}
corresponding to the number of years from the origin. The basis
functions are functions of this variable \eqn{t}. A year is
defined as \code{365.25} days, so \eqn{t} can not take only
integer values.



Consider the date as a numeric variable \eqn{t}, let \eqn{d} be
the degree.  The basis consist in

\enumerate{

\item the powers \eqn{t^0 = 1}, \eqn{t^1}, \eqn{\dots}{...},
\eqn{t^d} where \eqn{d} is the given degree.

\item the truncated powers \eqn{(t - \tau_k)^d_+} where \eqn{\tau_k}
is the break with number \eqn{k} and \eqn{x_+} is the positive
part of \eqn{x} (so \eqn{x_+ = 0} when \eqn{x < 0}).

}

The power zero will be discarded when \code{constant} is
\code{FALSE}.
}
\note{
When the argument \code{breaks} has length zero, the
function returns a polynomial basis with the constant possibly
removed.
}
\examples{

date <- as.Date(sprintf("\%4d-01-01", 1921:2020)) 
X1 <- breaksX(date = date, breaks = c("1970-01-01", "1990-01-01"))
plot(X1) 
}
\author{
Yves Deville
}
