% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{bs}
\alias{bs}
\title{Bootstrap an Object}
\usage{
bs(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Further arguments for methods.}
}
\value{
Bootstrap results
}
\description{
Bootstrap an object representing a fitted model.
}
