% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEV.R
\name{bs.TVGEV}
\alias{bs.TVGEV}
\title{Bootstrap for a \code{TVGEV} Model}
\usage{
\method{bs}{TVGEV}(object, R = 100, type = c("param", "NP"),
  estim = "optim", parallel = FALSE, ...)
}
\arguments{
\item{object}{A \code{TVGEV} object.}

\item{R}{Target number of bootstrap samples.}

\item{type}{Character. The value \code{"param"} leads to a
parametric bootstrap: for each bootstrap sample, a response vector
drawn from the estimated GEV distribution as described in
\code{object}; a MLE is performed to find the bootstraped
parameter vector. When \code{type} is \code{"NP"}, the generalised
residuals of the objects are resampled and are back-transformed to
simulated observations; a MLE is performed to to find the
bootstraped parameter vector as in the parametric case.}

\item{estim}{Argument passed to \code{MLE}}

\item{parallel}{Logical. If \code{TRUE} the loop over the bootstrap
sample is performed using \code{foreach} with \code{\%do par\%}}

\item{...}{Further arguments passed to \code{MLE}.}
}
\value{
A matrix with one row for each bootstrap sample, containing
the parameters.
}
\description{
Bootstrap for a \code{TVGEV} Model
}
\note{
The simulated response does not cover the observations where
the original response was NA.
}
\section{Caution}{
 for some bootstrap samples, the MLE may fail in
which case the coefficients will be set to \code{NA}.
}

\examples{

example(TVGEV)
bs <- bs.TVGEV(res2, R = 50, estim = "nloptr")

\dontrun{
   library(parallel)
   library(doParallel)
   nc <- detectCores()
   cl <- makeCluster(nc)
   registerDoParallel(cl)

   ## findings: with 'nloptr', less than 1\% of the optimisations
   ## diverges, while more than 10\% diverged with 'optim'.
   te <- system.time(bsp <- bs.TVGEV(res2, R = 5000, estim = "nloptr",
                                     parallel = TRUE))
   stopCluster(cl)
}


}
