% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bts.R
\name{bts}
\alias{bts}
\title{Create a \code{bts} Object Representing a Block Time-Series}
\usage{
bts(x, dateFrom, date)
}
\arguments{
\item{x}{A matrix with the time-series components as its columns.}

\item{dateFrom}{A beginning date used when \code{date} is
missing. Then a date vector}

\item{date}{A vector that can be coerced to the class
\code{"Date"} with length \code{nrow(x)}.}
}
\value{
An object with class \code{"bts"} inheriting from
\code{"matrix"}. This object has a \code{"date"} attribute which
is used to identify the beginning of the blocks.
}
\description{
Create a \code{bts} object representing a block time-series.
}
\details{
This structure is intended to represent \emph{slowly varying}
time-series only. So we can plot the column of \code{bts} object
against the date vector given as attribute. Since a date
represents the beginning of a block, each polyline (or "curve")
actually shifted by an half-block.
}
\note{
Some matrix operations do not make sense for \code{bts}
objects. A \code{"bts"} is transformed in an ordinary numeric
matrix by simply using \code{unclass}, see examples.
}
\examples{
x <- matrix(1:300, ncol = 10)
myBts1 <- bts(x, dateFrom = "2010-01-01")
## plot method
plot(myBts1)
mat <- unclass(myBts1)
class(mat)
x <- matrix(150 + 150 * runif(300), ncol = 100)
## with more than 12 columns, the legend is no longer shown
myBts2 <- bts(x, date = c("2040-01-01", "2041-01-01", "2042-01-01"))
plot(myBts2)
}
