% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{cdf}
\alias{cdf}
\title{Marginal Cumulative Distribution Functions}
\usage{
cdf(x, ...)
}
\arguments{
\item{x}{An object reprensenting a fitted model.}

\item{...}{Further arguments for methods.}
}
\value{
A structure containing discretized versions of the
marginal distribution functions for the model \code{x}.
}
\description{
Marginal cumulative distribution functions.
}
