% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantMax.R
\name{cdfMaxFun}
\alias{cdfMaxFun}
\title{Distribution Function for a Random Maximum}
\usage{
cdfMaxFun(object, ...)
}
\arguments{
\item{object}{An object that can be used to compute/predict the
distribution of the random maximum.}

\item{...}{Further arguments for methods.}
}
\value{
A function (or closure) that can be used to compute the
    probability of non-exceedance for arbitraty given
    quantiles.
}
\description{
Distribution Function for a Random Maximum
}
