% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantMax.R
\name{cdfMaxFun.TVGEV}
\alias{cdfMaxFun.TVGEV}
\title{Distribution Function for the Random Maximum on a given
    Period}
\usage{
cdfMaxFun.TVGEV(object, date = NULL, psi = NULL, theta = NULL)
}
\arguments{
\item{object}{An object with class \code{"TVGEV"}.}

\item{date}{An object that can be coerced to the class
\code{"Date"}. If not provided this will be taken as the date
attached to \code{object}.}

\item{psi}{An optional vector of coefficients for
\code{object}. By default the ML estimate as returned by
\code{coef(object)} is used.}
}
\value{
A function (more precisely, a closure). This function has
    a single formal argument \code{q} representing a quantile, and
    it returns the corresponding probability
    \eqn{\text{Pr}\{M^\star \leq q \}}{Pr[MStar <= q]}.
}
\description{
Given an object with class \code{"TVGEV"} and a collection of time
blocks (or period) defined by \code{date}, the distribution of the
random maximum over the period \eqn{M^\star := \max_b Y_b}{MStar =
max_b Y_b} is known. The corresponding distribution function can
be obtained.
}
\section{Caution}{
 When \code{theta} is given \code{model} is not
    used. The distribution function is simply that of the maximum
    of independent r.vs following GEV distributions with their
    parameters given by (the rows of) \code{theta}.
}

\examples{
df <- within(TXMax_Dijon, Date <- as.Date(sprintf("\%4d-01-01", Year)))
## fit a TVGEV model. Only the location parameter is TV.
res1 <- TVGEV(data = df, response = "TXMax", date = "Date",
              design = breaksX(date = Date, breaks = "1970-01-01", degree = 1),
              loc = ~ t1 + t1_1970)
cdf <- cdfMaxFun.TVGEV(res1)
cdf(c(39.0, 41.0))

## a 'new period'
date <- as.Date(sprintf("\%4d-01-01", 2025:2055))
cdfNew <- cdfMaxFun.TVGEV(res1, date = date)
cdfNew(c(39.0, 41.0))

}
\seealso{
\code{\link{quantileMaxFun.TVGEV}} for the corresponding
    quantile function (or closure).
}
