% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFuns.R
\name{check_confint}
\alias{check_confint}
\title{Check the confint Method for \code{TVGEV} Objects}
\usage{
check_confint(object, ref, level = 0.95, which = "shape", method = "proflik")
}
\arguments{
\item{object}{An object with class \code{"TVGEV"} representing a
stationary model.}

\item{ref}{An object with class \code{"gev.fit"} created by the
function \code{gev.fit} of the \pkg{ismev} package.}

\item{level}{Confidence level.}

\item{which}{Not used yet.}

\item{method}{See \code{\link{confint.TVGEV}}. For now, only
\code{"proflik"} is possible.}
}
\value{
Nothing is returned: on the plot built by
\code{ismev::prof.xi}, two thick vertical lines are added, showing
the confidence limits found by \code{confint.TVGEV} with
\code{method = "proflik"}. Each vertical line should cut
the profiled log-likelihood curve at one of its two intersections
with the horizontal line.
}
\description{
Check the \code{confint method}.
}
\details{
Confidence intervals on the parameters of \code{object} are
computed using the \code{confint} method for the class
\code{"TVGEV"}. The confidence interval for the shape parameter is
compared to that given by the \code{ismev::prof.xi} function.
}
\examples{
data(portpirie)
df <- portpirie
df <- within(df, date <- as.Date(sprintf("\%4d-01-01", Year)))
fitNSGEV <- try(TVGEV(data = df,
                       response = "SeaLevel", date = "date"))
fitismev <- gev.fit(xdat = df$SeaLev)
check_confint(fitNSGEV, ref = fitismev)

}
\seealso{
\code{\link{confint.TVGEV}}.
}
