% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFuns.R
\name{check_predict}
\alias{check_predict}
\title{Check the Predict Method for \code{TVGEV} Objects}
\usage{
check_predict(
  object,
  ref,
  level = 0.95,
  which = "confint",
  newdate = as.Date("2020-02-01"),
  confintMethod = "proflik"
)
}
\arguments{
\item{object}{A \code{TVGEV} package.}

\item{ref}{An object with class \code{"gev.fit"} created by the
function \code{gev.fit} of the \pkg{ismev} package.}

\item{level}{Confidence level.}

\item{which}{Not used yet.}

\item{newdate}{See \code{\link{predict.TVGEV}}.}

\item{confintMethod}{See \code{\link{predict.TVGEV}}.}
}
\value{
Nothing. For each return period computed by
\code{predict.TVGEV}, the plot built by \code{ismev::gev.prof} is
shown. Two vertical lines are added, showing
the confidence limits found by \code{predict.TVGEV} with
\code{confintMethod = "proflik"}. Each vertical line should cut
the profiled log-likelihood curve at one of its two intersections
with the horizontal line.
}
\description{
Check the \code{predict method}.
}
\details{
The comparison with the \pkg{ismev} package is only completed for
stationary models because it is difficult otherwise to get
profile-likelihood confidence intevals on Return Levels.
}
\note{
\code{ismev::gev.prof} does not return the confidence limits
which must be evaluated by eye from the displayed graph.
}
\examples{
data(portpirie)
df <- portpirie
df <- within(df, date <- as.Date(sprintf("\%4d-01-01", Year)))
fitNSGEV <- try(TVGEV(data = df,
                       response = "SeaLevel", date = "date"))
fitismev <- gev.fit(xdat = df$SeaLev)
check_predict(fitNSGEV, ref = fitismev)

}
\seealso{
\code{\link{predict.TVGEV}}.
}
