% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{checkNewdata.TVGEV}
\alias{checkNewdata.TVGEV}
\title{Check New Data for Prediction}
\usage{
checkNewdata.TVGEV(object, newdata = NULL, design = FALSE)
}
\arguments{
\item{object}{A \code{TSGEV} object.}

\item{newdata}{An atomic vector or a data frame providing the data
for a prediction.}

\item{design}{Not used yet.}
}
\value{
A data frame with the variables required for prediction.
    This must always contain a column with is name given by
    \code{object$date} and with class \code{"Date"}. If
    \code{object} contains \code{TSVars}, the corresponding
    variables must be provided as well, but with no check on their
    class.
}
\description{
Check New Data for Prediction
}
\keyword{internal}
