% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{d2negLogLikFun}
\alias{d2negLogLikFun}
\title{Hessian of the Negative Log-Likelihood of a \code{TVGEV} Object}
\usage{
d2negLogLikFun(psi, object)
}
\arguments{
\item{psi}{Value of the parameter vetor.}

\item{object}{A \code{TVGEV} object.}
}
\value{
The Hessian matrix of the negative log-likelihood. When
\code{psi} is taken as the estimate of the model given in
\code{object}, the matrix should be positive semidefinite.
}
\description{
Exact Hessian of the negative log-likelihood function of a
\code{TVGEV} object.
}
\details{
The Hessian matrix computation relies on the exact second order
derivatives of the GEV density w.r.t the vector
\eqn{\boldsymbol{\theta}}{\theta} of GEV parameters, see
\code{\link[nieve]{dGEV}}. The derivatives w.r.t. the vector of model
parameters \eqn{\boldsymbol{psi}}{\psi} are computed by chain rule
using the linear inverse link.
}
\note{
For now, only a fairly limited use is made of the Hessian,
hence we do not strive to optimise the computation. Note that the
Hessian used in the \code{\link{TVGEV}} estimation function
(creator of \code{TVGEV} objects) and providing the covariance
matrix \code{vcov} by inversion is a \emph{numeric} one for now.
}
\examples{
\dontrun{
   example(TVGEV)
   H1 <- d2negLogLikFun(res1$estimate, object = res1)
   H1 - res1$hessian
   H2 <- d2negLogLikFun(res2$estimate, object = res2)
   H2 - res2$hessian
}
}
