% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSGEV.R
\name{density.NSGEV}
\alias{density.NSGEV}
\title{Compute NSGEV Densities}
\usage{
\method{density}{NSGEV}(x, xValue = NULL, data = NULL, psi = NULL,
  log = FALSE, ...)
}
\arguments{
\item{x}{A \code{NSGEV} object.  Note that the name of this formal
is imposed by the S3 generic, it would otherwise probably have
been chosen as 'model'.}

\item{xValue}{Vector of quantiles at which the GEV densities will be
evaluated. By default, a grid of value is found with coverage
probability \code{> 0.001} for each observation.}

\item{data}{A data frame with the covariates. If \code{NULL}, the
data frama affached to the model given in \code{x} is used.}

\item{psi}{Vector of model coefficients.}

\item{log}{Logical. If \code{TRUE} the log-density is returned.}

\item{...}{Not used yet.}
}
\value{
A matrix of density values, with one row for each observation
of \code{data} and one column by quantile.
}
\description{
Compute NSGEV densities.
}
\examples{
df <- data.frame(t = 1:10)
## model structure
ns <- NSGEV(formulas = list("loc" = ~ alpha + beta * t,
                            "scale" = ~ delta,
                            "shape" = ~ xi),
             data = df)
psi <- c("alpha" = 1, "beta" = 0.1, "delta" = 0.6, "xi" = 0.06)
d <- density(ns, psi = psi) 
matplot(x = attr(d, "x"), y = t(d), type = "l")

}
\author{
Yves Deville
}
