% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTVGEV.R
\name{density.TVGEV}
\alias{density.TVGEV}
\alias{cdf.TVGEV}
\title{Compute TVGEV Densities or Cumulative Distribution
Functions}
\usage{
\method{density}{TVGEV}(x, xValue = NULL, date = NULL, psi = NULL, log = FALSE, ...)

\method{cdf}{TVGEV}(x, qValue = NULL, date = NULL, psi = NULL, log = FALSE, ...)
}
\arguments{
\item{x}{A \code{TVGEV} object.}

\item{xValue}{Vector of quantiles at which the GEV densities will
be evaluated. By default, a grid of value is found with coverage
probability > 0.001 for each observation.}

\item{date}{An object that can be coerced to the
class\code{"Date"} giving the date of the blocks for which the
density will be evaluated. By default, three dates are selected
in the date vector attached to \code{x}.}

\item{psi}{Vector of model coefficients. By default, the vector of
estimated coefficients in \code{x} is used.}

\item{log}{Logical. If TRUE the log-density is returned.}

\item{...}{Other arguments to be passed to
\code{\link[nieve]{dGEV}} or \code{\link[nieve]{pGEV}}.}

\item{qValue}{xValue Vector of probabilities at which the GEV
densities will be evaluated. By default, a grid of value from
\code{0.0} to \code{1.0} is used. for each observation.}
}
\value{
An oject with class \code{"bfts"}. This is mainly a matrix
with one row by date. Rather than providing a time-plot of each
column (as would be done for a \code{"bts"} object), the
\code{plot} method plots the density or cdf function for a small
number of dates. Unless the number of dates is \code{1}, the
functions are plotted with the x-y axes flipped in order to enhance
the time-varying feature of the model.
}
\description{
Compute TVGEV densities or Cumulative Distribution Functions.
}
\examples{
example(TVGEV)
d <- density(res1, date = c("2000-01-01", "2020-01-01", "2040-01-01"))
F <- cdf(res1, date = c("2000-01-01", "2020-01-01", "2040-01-01"))
plot(d, fill = TRUE)
plot(F)

}
\seealso{
\code{\link[nieve]{GEV}} for the density and cdf of the GEV
distribution, \code{\link{plot.predict.TVGEV}} for the Return Level
plot.
}
