% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{distLines}
\alias{distLines}
\title{Distance Between Two Lines}
\usage{
distLines(x1, x2)
}
\arguments{
\item{x1}{A numeric vector with positive norm.}

\item{x2}{A numeric vector with positive norm and with the same
length as \code{x1}.}
}
\value{
The distance.
}
\description{
Distance between two lines.
}
\details{
The vectors \code{x1} \code{x1} are scaled to have unit
Euclidean norm and to have the same direction. The distance
between the lines is simply the distance between the unit vectors
on the unit sphere.
}
