% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantMax.R
\name{dMax.TVGEV}
\alias{dMax.TVGEV}
\title{Value of the Probability Density of the Maximum over a
    Period of Time}
\usage{
dMax.TVGEV(
  object,
  x,
  date = NULL,
  psi = NULL,
  deriv = FALSE,
  derx = FALSE,
  ...
)
}
\arguments{
\item{object}{A \code{TVGEV} object.}

\item{x}{A vector of values at which the density is to be
evaluated.}

\item{date}{A vector that can be coerced to the class
\code{"Date"} which defines the period of time.}

\item{psi}{Optional vector of parameters. By default, the
estimated vector of parameters \code{coef(object)} is used.}

\item{deriv}{Logical. If \code{TRUE} the gradient of the density
will be computed and returned as the \code{"gradient"}
attribute of the result.}

\item{derx}{Logical, If \code{TRUE} the derivative of the density
w.r.t. \code{x} will be computed and returned as the
\code{"derx"} attribute of the result.}

\item{...}{Not used yet.}
}
\value{
A vector of values for the probability density. Beside the
    gradient (w.r.t. the parameters) returned as the attribute
    named \code{"gradient"}, the derivative w.r.t. \code{x} is
    also returned as the attribute \code{"derx"} of the result.
}
\description{
Compute the probability density of the maximum over a period of
time.
}
