% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantMax.R
\name{.qMax.TVGEV}
\alias{.qMax.TVGEV}
\title{Upper Bound for the Quantile of the Random Maximum using a
    \code{"TVGEV"} object.}
\usage{
.qMax.TVGEV(theta, p)
}
\arguments{
\item{theta}{a numeric matrix containing the GEV "theta"
parameters. It must have 3 columns corresponding for
\emph{location} \eqn{\mu}, \emph{scale} \eqn{\sigma} and
\emph{shape} \eqn{\xi} is that order.}

\item{p}{A probability.}
}
\description{
Find an upper bound for the quantile of the random maximum. The
distribution function of the maximum is the product of the
marginal (GEV) distributions functions for the blocks over which
the maximum is defined. By selecting the largest GEV shape, the
largest scale and the largest location, we get a GEV distribution
which is larger in stochastic order than all the marginal GEV
distributions. We can take the bound as the quantile of this GEV
distribution corresponding to the probability
\eqn{p^{1/n^\star}}{p^(1/ nStar)}, where \eqn{n^\star}{nStar} is
the number of blocks.
}
