% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influenceTVGEV.R
\name{influence.TVGEV}
\alias{influence.TVGEV}
\title{Diagnostic of Influence for \code{TVGEV} Objects}
\usage{
\method{influence}{TVGEV}(
  model,
  what = function(model) coef(model)["xi_0"],
  which = "min",
  how = "replace",
  ...
)
}
\arguments{
\item{model}{A \code{TVGEV} object.}

\item{what}{A function with one argument \code{object} defining or
extracting the (numeric) quantity for which the influence is
to be analysed. So, \code{object} must be a \code{TVGEV}
object. See \bold{Examples}. This function can return numeric
vector in which case several influence curves will be
computed. It is then wise to make the function return a vector
with named elements.}

\item{which}{Identifies the observation which influence will be
analysed. Can be: an integer giving the observation, the
character \code{"min"}, the character \code{"max"}, a
character that can be coerced to a date, or an object with
class \code{"Date"} that identifies the observation.  This
argument must define an observation existing in the data frame
used to create the object. So it can not be used e.g. to
define a future date.}

\item{how}{Character specifying the type of (finite-sample)
influence curve to use. For now only the only possibility is
\code{"replace"}, see Section 2.1 of the book by Hampel et al
(1996).  The chosen observation is replaced by a value \eqn{x}
and the quantity of interest is computed for each value of
\eqn{x}.}

\item{...}{Not used.}
}
\value{
A named list of vectors. The element \code{obsValue}
    contains the candidate values for the chosen observation and
    \code{statValue} contains the corresponding values of the
    specific statistic defined with \code{what}.
}
\description{
This function provides some indications on the influence of a
specific observation used in the fit of a \code{TVGEV} object. For
several candidate values of the chosen observation, a \code{TVGEV}
is re-fitted with the observed value replaced by the candidate
value, and the quantity of interest as defined by \code{what} is
computed.
}
\details{
By suitably defining the function \code{what}, one can trace e.g.,
the estimate of a quantile with given probability, the
corresponding upper confidence limit, the upper end-point of the
distribution, ... and more.
}
\note{
For \code{TVGEV} models with a positive GEV shape, the
    smallest observation may have a strong influence on the
    result, even if the model is stationary.
}
\examples{
example(TVGEV)
influ <- influence(res2, which = "min")
autoplot(influ)
influ <- influence(res2, which = as.Date("2015-01-01"))
autoplot(influ)
RL_2050 <- function(model) {
   c("RL_2050(100)" = quantile(model, prob = 0.99, date = "2050-01-01")[1])
}
influence(res2, what = RL_2050) |> autoplot()
influence(res2, what = RL_2050, which = "2003-01-01") |> autoplot()
RLs_2050 <- function(model) {
   res <- c(quantile(model, prob = 0.99, date = "2050-01-01")[1],
            quantile(model, prob = 0.999, date = "2050-01-01")[1])
   names(res) <- paste0("RL", c(100, 1000))
   res
}
influ <- influence(res2, what = RLs_2050, which = "2003-01-01")
autoplot(influ)
}
\references{
Hampel, F.R., Ronchetti, E.M., Rousseeuw, P.J. and Stahel
  W.A. (1996) \emph{Robust Statistics: The Approach Based on
  Influence Functions}. Wiley.
}
