% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEV.R
\name{logLik.TVGEV}
\alias{logLik.TVGEV}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{TVGEV}(object, ...)
}
\arguments{
\item{object}{An object with class \code{"TVGEV"} representing a
time-varying GEV model.}

\item{...}{Not used.}
}
\value{
The maximized log-Likelihood.
}
\description{
Extract log-likelihood.
}
