% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTVGEV.R
\name{mean.TVGEV}
\alias{mean.TVGEV}
\alias{moment.TVGEV}
\title{Expectation or Moments of a \code{TVGEV} Model Object}
\usage{
\method{mean}{TVGEV}(x, date = NULL, psi = NULL, ...)

\method{moment}{TVGEV}(x, which = "variance", date = NULL,
  psi = NULL, ...)
}
\arguments{
\item{x}{An \code{TVGEV} object.}

\item{date}{An object that can be coerced to the \code{"Date"}
class. By default the date in \code{x} is used.}

\item{psi}{Vector of parameters. By default the vector of the
parameters for the model \code{x} is used.}

\item{...}{Not used yet.}

\item{which}{Description of the moment. For now only the value
\code{"variance"} is accepted, but a description for the 3-rd and 4-th
order moments should be accepted in the future.}
}
\value{
An object with class \code{"bts"} representing a block
time series with its value set to the expectation or the moment of
the GEV distribution.
}
\description{
Expectation or moments of a \code{TVGEV} model.
}
\examples{
example(TVGEV)
plot(mean(res1))
}
\seealso{
\code{\link{quantile.TVGEV}}
to compute quantiles of the time-varying distribution.
}
