% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{modelMatrices}
\alias{modelMatrices}
\title{Construct or Extract Design Matrices}
\usage{
modelMatrices(object, ...)
}
\arguments{
\item{object}{An object representing a model.}

\item{...}{Further arguments for methods.}
}
\value{
A lise of desing matrices, usually one by parameter of the
    distribution.
}
\description{
Construct or extract the design (or model) matrices.
}
\seealso{
\code{\link[stats]{model.matrix}}.
}
