% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEV.R
\name{modelMatrices.TSGEV}
\alias{modelMatrices.TSGEV}
\title{Build Model Matrices}
\usage{
modelMatrices.TSGEV(object, date = NULL)
}
\arguments{
\item{object}{An object with S3 class \code{"TSGEV"}.}

\item{date}{An object with class \code{"Date"} or an object that
can be coerced to this class.}
}
\value{
A list with the follwing elements

\item{dfAll}{

A data frame with all the variables required.

}
\item{X}{

A list with three elements named \code{"loc"}, \code{"scale"} and
\code{"shape"}. Each element is the corresponding model matrix in
the \code{lm} meaning. Each matrix can be multiplied by the
corresponding coefficients to get the prediction for the
corresponding parameter.

}
}
\description{
Build model matrices.
}
\details{
These matrices are needed when the model is build (and usually is
estimated) or when a prediction is required.
}
\seealso{
The function \code{\link[stats]{model.matrix}} used by
\code{\link[stats]{lm}}.
}
