% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEV.R
\name{modelMatrices.TVGEV}
\alias{modelMatrices.TVGEV}
\title{Build or Extract Model Matrices}
\usage{
\method{modelMatrices}{TVGEV}(object, date = NULL)
}
\arguments{
\item{object}{An object with S3 class \code{"TVGEV"}.}

\item{date}{An object with class \code{"Date"} or an object that
can be coerced to this class.}
}
\value{
A list with the follwing elements

\item{dfAll}{

A data frame with all the variables required.

}
\item{X}{

A list with three elements named \code{"loc"}, \code{"scale"} and
\code{"shape"}. Each element is the corresponding model matrix in
the \code{lm} meaning. Each matrix can be multiplied by the
corresponding coefficients to get the prediction for the
corresponding parameter.

}
}
\description{
Build or extract model matrices. The three matrices corresponding
to the GEV parameters (loc, scale and shape) are returned into a
named list of matrices. For each matrix the rows correspond to the
observations of the model and the columns relate to the
corresponding formula.
}
\details{
These matrices are needed when the model is build (and usually is
estimated) or when a prediction is required.
}
\seealso{
The function \code{\link[stats]{model.matrix}} used by
\code{\link[stats]{lm}}.
}
