% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{moment}
\alias{moment}
\title{Moments of Marginal Distributions}
\usage{
moment(x, which, ...)
}
\arguments{
\item{x}{An object representing a model.}

\item{which}{The order of the moment.}

\item{...}{Further arguments for methods.}
}
\value{
A vector of moments.
}
\description{
Compute the moments of the marginal distributions attached to a
model object.
}
