% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSGEV.R
\name{negLogLik}
\alias{negLogLik}
\title{Evaluate the negative log-likelihood of a NSGEV model}
\usage{
negLogLik(psi, model, data = NULL, y = NULL, deriv = TRUE, checkNames = TRUE)
}
\arguments{
\item{psi}{Vector of NSGEV parameters.}

\item{model}{The NSGEV model.}

\item{data}{A data frame with the covariates to be used. By
default, the data attached to \code{model} is udes.}

\item{y}{The response as in \code{NSGEV}. By default the response
attached to the model is used.}

\item{deriv}{Logical. When \code{TRUE}, the gradient
\eqn{-\partial \ell/\partial \boldsymbol{\psi}}{-d log L/ d \psi}
of the negative log-likelihood w.r.t. the parameters of the model
will be returned. In this case, a list will be returned.}

\item{checkNames}{Logical. If \code{TRUE}, the names of the vector
\code{psi} are checked against the parnames of the model.}
}
\value{
The negative log-likelihood (to be minimized) or a list
with two elements corresponding to the value of the log-likelihood
and its gradient.
}
\description{
Evaluate the negative log-likelihood of a NSGEV model.
}
\note{
The returned value of this function when \code{deriv} is \code{TRUE}
is designed for use with the \code{nloptr} package.
}
\section{Caution}{
 The class of the result may be changed in future
versions.
}

\author{
Yves Deville
}
