\name{NSGEV-package}
\alias{NSGEV-package}
\docType{package}
\title{
\packageTitle{NSGEV}
}
\description{
\packageDescription{NSGEV}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NSGEV}
\packageIndices{NSGEV}

This package. As a major difference with comparable functions found in some other EV packages, 
the model parameters can be used in several of the GEV parameters \code{loc}, \code{scale} and \code{shape}.
A typical example is when a scale proportional to the location needs to be specified.
}
\author{
\packageAuthor{NSGEV}

Maintainer: \packageMaintainer{NSGEV}
}
\references{
  Coles, S. (2001) \emph{Introdution to Statistical Modelling of Extreme Values}, Springer.
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{



}
