% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{opacity}
\alias{opacity}
\title{Choose an Opacity Level}
\usage{
opacity(n)
}
\arguments{
\item{n}{Vector of integers.}
}
\value{
Vector of opacities "alpha". Value 0 is for a fully
transparent colour, value 1 is for a fully opaque colour.
}
\description{
Choose an opacity level.
}
