% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEV.R
\name{parIni.TVGEV}
\alias{parIni.TVGEV}
\title{Initial Parameter Estimates}
\usage{
parIni.TVGEV(object, y = NULL)
}
\arguments{
\item{object}{A (possibly incomplete) \code{TVGEV} object.}

\item{y}{The response, by defaut the response attached to the
model.}
}
\value{
A vector of initial values.
}
\description{
Initial Parameter Estimates for a \code{TVGEV} model.
}
\details{
The model parameters that are inverse-linked to the location
parameter are found by using a linear model with a constant error
variance. The initial value of the shape parameter(s) being set to
zero, which correspond to Gumbel margins, we know that the
expectation is approximately \eqn{\mu + 0.577 \sigma} while
the standard deviation is \eqn{\sigma \sqrt{\pi} / 6}
}
