% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{parNames.default}
\alias{parNames.default}
\title{Names of the parameters of a model}
\usage{
\method{parNames}{default}(object, ...)
}
\arguments{
\item{object}{A parametric statistical model for which the
parameters must have names.}

\item{...}{Not used yet.}
}
\value{
The character vector of the names of the parameters of the
model.
}
\description{
Names of the parameters of a model.
}
