% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allMethods.R
\name{parNames}
\alias{parNames}
\title{Names of the parameters of a statistical model}
\usage{
parNames(object, ...)
}
\arguments{
\item{object}{A parameteric statistical model for which the
parameters must have names.}

\item{...}{Not used yet.}
}
\value{
The character vector of the names of the parameters of the
model.
}
\description{
Names of the parameters of a model.
}

