% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bts.R
\name{plot.bfts}
\alias{plot.bfts}
\title{Plot for Block Functional Time Series \code{bfts} Objects}
\usage{
\method{plot}{bfts}(x, y, gg = TRUE, col = "orangered", fill = NA,
  ...)
}
\arguments{
\item{x}{A \code{bfts} object such as returned by the \code{density} and
\code{cdf} methods for \code{TVGEV} objects.}

\item{y}{Not used yet.}

\item{gg}{Logical. If \code{TRUE}, a \code{ggplot} object is shown
and returned.}

\item{col}{Colour for the line and fill.}

\item{fill}{If non \code{NA}, the region between the curve and the
axis will be filled as is quite usual with probability densities.}

\item{...}{Not used yet.}
}
\value{
A \code{ggplot} object or nothing.
}
\description{
Plot for Block Functional Time Series \code{bfts} objects.
}
\note{
This function is mainly intended to be used for density
functions or cdf. In the former case, \code{fill = TRUE} may be
preferred.
}
