% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influenceTVGEV.R
\name{plot.influence.TVGEV}
\alias{plot.influence.TVGEV}
\title{Plot Method for \code{influence.TVGEV} Object}
\usage{
\method{plot}{influence.TVGEV}(x, y = NULL, ...)
}
\arguments{
\item{x}{A \code{TVGEV} object.}

\item{y}{Not used.}

\item{...}{Not used.}
}
\value{
Nothing.
}
\description{
Plot Method for \code{influence.TVGEV} Object
}
\note{
This method \code{autoplot}
}
\seealso{
\code{\link{autoplot.TVGEV}}.
}
