% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allNSGEVMethods.R
\name{plot.NSGEV}
\alias{plot.NSGEV}
\title{Plot method for \code{NSGEV} object}
\usage{
\method{plot}{NSGEV}(
  x,
  y,
  which = 1,
  pLim = c(0.01, 0.99),
  showQuant = c(0.5, 0.95, 0.99),
  ...
)
}
\arguments{
\item{x}{A NSGEV object}

\item{y}{Not used.}

\item{which}{An integer. The value \code{1} will lead to a plot
with the observation number used for the \eqn{x} axis. With the
value \code{2} the \eqn{x} axis is the covariate, assuming that
there is only one covariate.}

\item{pLim}{A numeric vector with two probabilities used to set
the \code{ylim} parameter of \code{plot.default}. The model
quantiles for these probabilites are computed and added to the
range used to choose \code{ylim}.}

\item{showQuant}{Not used yet.}

\item{...}{Other arguments to be passed to methods.}
}
\value{
Nothing.
}
\description{
Plot method for \code{NSGEV} objects.
}
\examples{
example(NSGEV)
plot(ns1, which = 1)
plot(ns1, which = 2)

}
\author{
Yves Deville
}
