% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictTVGEV.R
\name{plot.predict.TVGEV}
\alias{plot.predict.TVGEV}
\title{Plot Predict Results for \code{TVGEV}}
\usage{
\method{plot}{predict.TVGEV}(x, y, gg = TRUE, ...)
}
\arguments{
\item{x}{An object with class \code{"predict.TVGEV"} as returned
by the \code{predict} method. This must be a data frame so the
\code{predict} method must be called without using \code{out} or
setting it to the default value \code{"data.frame"}.}

\item{y}{Not used.}

\item{gg}{Logical. If \code{TRUE} the plot is produced with the
\bold{ggplot2} package.}

\item{...}{Not used yet.}
}
\value{
An object with class \code{"gg"}. Can be used with the method
\code{plot}, or equivalently with \code{print}.
}
\description{
Plot predict Results for \code{TVGEV}
}
\note{
This function is intended to work with predictions computed
for a possibly large number of periods (to get smooth curves) but
with only a small number of dates, each apearing in a facet. So
the number of dates is limited to \code{6}.
}
\examples{
example(TVGEV)
pred <- predict(res2, newdate = c("1960-01-01", "2000-01-01", "2020-01-01"),
                level = c(0.70, 0.95), confintMethod = "delta")
## no longer needed
## class(pred) <- c("predict.TVGEV", "data.frame")
g <- plot(pred)
}
