% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residTVGEV.R
\name{plot.resid.TVGEV}
\alias{plot.resid.TVGEV}
\title{Plot Residuals of a \code{TGVEV} model.}
\usage{
\method{plot}{resid.TVGEV}(x, y = NULL, ...)
}
\arguments{
\item{x}{An object with class \code{"TVGEV"}.}

\item{y}{Not used yet.}

\item{...}{Further arguments to be passed to \code{plot}.}
}
\value{
Nothing.
}
\description{
Plot the residuals of a \code{TVGEV} model object against the
date.
}
\seealso{
\code{\link{residuals.TVGEV}}.
}
