% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantMax.R
\name{pMax.TVGEV}
\alias{pMax.TVGEV}
\title{Value of the Distribution Function of the Maximum over a
    Period of Time}
\usage{
pMax.TVGEV(
  object,
  q,
  date = NULL,
  psi = NULL,
  deriv = FALSE,
  hessian = FALSE,
  ...
)
}
\arguments{
\item{object}{A \code{TVGEV} object.}

\item{q}{A vector of "quantile" values at which the distribution
function is to be evaluated.}

\item{date}{A vector that can be coerced to the class
\code{"Date"} which defines the period of time.}

\item{psi}{Optional vector of parameters. By default, the
estimated vector of parameters \code{coef(object)} is used.}

\item{deriv}{Logical. If \code{TRUE} the gradient of the
probability distribution will be computed and returned as the
\code{"gradient"} attribute of the result.}

\item{hessian}{Logical. If \code{TRUE} the Hessian of the density
will be computed and returned as the \code{"hessian"}
attribute of the result.}

\item{...}{Not used yet.}
}
\value{
A vector of values for the distribution function. Beside
    the gradient (w.r.t. the parameters) returned as the attribute
    named \code{"gradient"}, the Hessian w.r.t. \code{x} can
    also be returned as the attribute \code{"hessian"} of the result.
}
\description{
Compute the distribution function of the maximum over a period of
time.
}
