% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{polynomX}
\alias{polynomX}
\title{Design Matrix for Polynomial Regression}
\usage{
polynomX(date, degree = 2, origin = NULL)
}
\arguments{
\item{date}{A vector with class \code{"Date"} or that can be
coerced to this class.}

\item{degree}{The maximal degree \eqn{d}. The basis contains
\eqn{d + 1} functions \eqn{t^0}, \eqn{t^1}, \dots, \eqn{t^d}.}

\item{origin}{Optional vector of length 1 with class \code{"Date"}
or character that can be coerced to this class. Gives the origin
for the transformation of dates into numeric values. The default
value is \code{1970-01-01}, see \code{\link{Date}}.}
}
\value{
A matrix having the value of the basis functions as its
columns. Each row corresponds to an element of the given
\code{date} vector. By convention, the columns are named
\code{Cst} then \code{t1}, \code{t2}, \dots.
}
\description{
Build a design matrix for polynomial regression.
}
\details{
The provided date is converted into a numeric variable \eqn{t}
corresponding to the number of years from the origin. The basis
functions are functions of this variable \eqn{t}. A year is
defined as \code{365.25} days, so \eqn{t} can not take only
integer values.
}
\note{
This function is intended to provide regressors for Time
Varying GEV models corresponding to \code{TVGEV} objects. Since
the response then usually corresponds to block maxima, the date
variable gives either the begining or the end of each block.
}
\section{Caution}{
 As a general rule, it is a good practice to
choose \code{origin} close to the centre of the time period. This
is especially important when a polynomial of degree \eqn{>= 2} is
used since problems of convergence are likely to occur otherwise.
}

\examples{

date <- seq(from = as.Date("1996-01-01"),
            to = as.Date("2016-12-31"), by = "years")

X <- polynomX(date)
}
\seealso{
\code{breaksX} for a basis of Truncated Power functions.
and \code{TVGEV} to for Time Varying GEV models.
}
