% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictTVGEV.R
\name{predict.TVGEV}
\alias{predict.TVGEV}
\title{Predict Method for \code{TVGEV} Objects}
\usage{
\method{predict}{TVGEV}(
  object,
  newdate = NULL,
  period = NULL,
  level = 0.95,
  confintMethod = c("delta", "none", "boot", "proflik"),
  out = c("data.frame", "array"),
  biasCorrect = FALSE,
  trace = 1L,
  ...
)
}
\arguments{
\item{object}{An object with S3 class \code{"TVGEV"}}

\item{newdate}{A vector with class \code{"Date"} or that can be
coerced to this class.}

\item{period}{Numeric vector of periods expressed as multiple of
the block duration. Usually the block duration is one year and a
value \code{100} will correspond to a 100-year return level, i.e.
to the GEV quantile with probability 0.99. The default value
allows the construction of an acceptable RL plot, but if a
smoother RL curve and/or Confidence Band is needed, more values
would be used.}

\item{level}{Numeric vector of confidence level(s).}

\item{confintMethod}{The method used to compute the confidence
intervals. See \bold{Details}.}

\item{out}{Type of output.}

\item{biasCorrect}{\code{Logical} used only when
\code{confintMethod} is \code{"boot"}. When \code{TRUE}, the RL
named \code{"Quantile"} is computed as the average of the RLs
obtained with the bootstraped parameters, and it will differ from
the RL computed with the estimated parameters in \code{object}.}

\item{trace}{Integer level of verbosity.}

\item{...}{Arguments to be passed to \code{bs}.}
}
\value{
A data frame or an array with the RL and confidence limits
for each combination of \emph{date}, \emph{period} and
\emph{confidence level}. So when \code{confintMethod} is not
\code{"none"} the three values \code{Quant}, \code{L} and \code{U}
are given for each combination (the first not depending on the
confidence level). The data frame is in 'long' format, with
 different rows for multiple levels.
}
\description{
Predict method for \code{TVGEV} objects.
}
\details{
Compute the Return Levels (RLs) for different periods and
different values of the time block, as well as Confidence
Intervals (CIs) for these. The results are not predictions in the
usual acceptance.  The CIs can be obtained by the usual
\emph{delta method}, provided that the approximate covariance for
the estimated parameters is found in \code{object}. They also can
be obtained by \emph{bootstrap}, using by default the bootstrap
distribution embeded in \code{object} if any or by computing it
else. The bootstrap can be parametric or non-parametric, see
\code{\link{bs.TVGEV}}. Finally, the \emph{profile-likelihood}
method can be used: the confidence limits for a RL are obtained by
maximising and minimising this RL under the constraint that the
log-likelihood is greater than a suitable value. This method
avoids the usual re-parameterisation of the model which can be
tedious for the general form of model allowed in \code{TVGEV}.
}
\note{
Despite of its positive sounding name, the "bias correction"
can have a negative impact with some Extreme Value models as used
here, especially for non-parametric bootstrap. In future versions,
the dots \code{...} might be used to control the profile-likelihood
as well as the bootstrap.
}
\section{Caution}{
 When \code{confintMethod} is set to
\code{"loglik"} the required computing time is huge because
several constrained optimisations are required. Since the marginal
distribution hence the RL vary only slowly in practice, an
unnecessary computing burden will result when many values of
\code{newdate} are used.
}

\examples{
example(TVGEV)
p1 <- predict(res2, confintMethod = "none")
p2 <- predict(res2, confintMethod = "delta")
p3 <- predict(res2, confintMethod = "boot")
p4 <- predict(res2, newdate = "2020-01-01", confintMethod = "proflik") 
}
\seealso{
The bootstrap method \code{\link{bs.TVGEV}} for
\code{TVGEV} objects, the formal arguments of which can be used
here when \code{confintMethod} is chosen as \code{"boot"}.
}
\author{
Yves Deville
}
