% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bts.R
\name{print.bts}
\alias{print.bts}
\title{Print Method for \code{bts} Objects Representing Block Time
Series}
\usage{
\method{print}{bts}(x, ...)
}
\arguments{
\item{x}{A \code{bts} object.}

\item{...}{Not used yet.}
}
\value{
Nothing.
}
\description{
Print method for \code{bts} objects representing Block Time Series.
}
\details{
While \code{bts} objects are inheriting from the \code{"matrix"}
class, their \code{print} method aims to show the date range in
compact format a dozen of lines with limited text width.
}
