% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profLik.R
\name{profLik}
\alias{profLik}
\title{Profile-Likelihood Inference Method}
\usage{
profLik(object, fun, ...)
}
\arguments{
\item{object}{An object representing a parametric model.}

\item{fun}{A numeric function of the vector of parameters.}

\item{...}{Further arguments for methods.}
}
\value{
The result, typically a numeric array.
}
\description{
Profile-likelihood inference method.
}
\author{
Yves Deville
}
