% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profLik.R
\name{profLik.TVGEV}
\alias{profLik.TVGEV}
\title{Profile-Likelihood Inference for \code{TVGEV} Objects}
\usage{
\method{profLik}{TVGEV}(object, fun, level = 0.95, deriv = TRUE,
  trace = 0, ...)
}
\arguments{
\item{object}{A \code{TSGEV} object.}

\item{fun}{A function of the parameter for which the
profile-likelihood will be carried over. This function must have
the arguments: \code{psi} for the vector of parameters and
\code{object} for the model object. If needed, a wrapper function
can be used use more arguments, see \bold{Details}.}

\item{level}{Level of confidence. Can be of length \code{> 1}.}

\item{deriv}{Logical. If \code{TRUE}, the function \code{fun} is
assumed to provide a gradient vector as an attribute named
\code{"gradient"} of the result. For now \code{deriv} can only be
\code{TRUE}.}

\item{trace}{Level of verbosity; \code{trace = 0} prints nothing.}

\item{...}{Not used yet.}
}
\value{
An array with the value of the function and the
corresponding Lower and Upper end-points for the given confidence
levels.
}
\description{
Profile-likelihood inference for \code{TVGEV} objects.
}
\details{
Compute the lower and upper end-points of a profile-likelihood
based confidence interval. The (apparently new) method used here
relies on maximising the function of interest, say
\eqn{\rho(\boldsymbol(\psi)}{\rho(\psi)}, under the constraint
that the log-likelihood is greater than the maximal log-likelihood
minus a positive quantity \eqn{\delta} depending on the confidence
level. This differs from the usual method which relies on a
univariate zero-finding for the profile-likelihood function, the
evaluation of which relies on a \eqn{p-1} dimensional
optimisation. As a major advantage, the new method does not require
a re-parameterisation of the model.
}
\examples{
df <- within(TXMax_Dijon, Date <- as.Date(sprintf("\%4d-01-01", Year)))
## fit a TVGEV model with constant parameters.
res1 <- TVGEV(data = df, response = "TXMax", date = "Date",
              estim = "nloptr")
## define a function of the parameter vector
myfun <- function(psi, object, deriv = TRUE) {
    res <- psi[1]
    grad <- rep(0.0, object$p)
    grad[1] <- 1
    attr(res, "gradient") <- grad
    res
}

profLik.TVGEV(object = res1, fun = myfun, deriv = TRUE)

}
\references{
Deville Y. (2017) "Profile-likelihood using constrained
optimisation". Unpublished Tech. Report.
}
\author{
Yves Deville
}
