% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSGEV.R
\name{psi2theta.NSGEV}
\alias{psi2theta.NSGEV}
\title{Compute the Matrix of GEV Parameters from the Vector of
Model Parameters 'psi' and the Data}
\usage{
\method{psi2theta}{NSGEV}(model, psi, data = NULL, deriv = TRUE, checkNames = TRUE, ...)
}
\arguments{
\item{model}{The GEV model: an object with class \code{"NSGEV"}.}

\item{psi}{The vector of model parameters.}

\item{data}{The data frame of covariates. By default, the data
frame attached to the model is used.}

\item{deriv}{Logical. If true, the jacobian is computed and
returned as an attribute of the results. It is an array with
dimension \code{c(n, p, 3L)}}

\item{checkNames}{Logical. If \code{TRUE}, the names of the vector
\code{psi} are checked against the parnames of the model.}

\item{...}{Not used yet.}
}
\value{
A vector of parameters for the \code{NSGEV} model.
}
\description{
Compute the matrix of GEV parameters from the vector of model
parameters 'psi' and the Data.
}
\examples{
df <- data.frame(t = 1:10)
fit <- NSGEV(formulas = list("loc" = ~ alpha + beta * t, "scale" = ~ delta, "shape" = ~ xi),
data = df)
df.new <- data.frame(t = 11:20)
psi <- c("alpha" = 1, "beta" = 0.01, "delta" = 0.6, "xi" = 0.06)
theta.new <- psi2theta(model = fit, psi = psi, data = df.new)
matplot(df.new$t, theta.new, type = "b")
}
\author{
Yves Deville
}
