% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{psi2theta}
\alias{psi2theta}
\title{Compute GEV Parameters from Model Parameters}
\usage{
psi2theta(model, psi, ...)
}
\arguments{
\item{model}{An object representing a model with GEV margins.}

\item{psi}{A vector of parameters for the model.}

\item{...}{Not used yet.}
}
\value{
A vector or matrix of GEV parameters.
}
\description{
Compute vector(s) of GEV Parameters \code{theta} from vector(s)
of model parameters \code{psi}.
}
