% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEV.R
\name{psi2theta.TVGEV}
\alias{psi2theta.TVGEV}
\title{Compute the Matrix of GEV Parameters from the Vector of
Model Parameters 'psi' and the Data}
\usage{
\method{psi2theta}{TVGEV}(
  model,
  psi = NULL,
  date = NULL,
  deriv = FALSE,
  checkNames = TRUE,
  ...
)
}
\arguments{
\item{model}{The \code{TVGEV} object.}

\item{psi}{A vector of parameters for \code{model}. By default the
vector of estimates in \code{psi} will be used.}

\item{date}{A vector that can be coerced to the \code{"Date"}
class. Each element will correspond to a row of the matrix of GEV
parameters.}

\item{deriv}{Logical. If \code{TRUE} the returned object has an
attribute names \code{"gradient"} which is the gradient
array.}

\item{checkNames}{Logical. If \code{TRUE} it will be checked that
the names of \code{psi} match the \code{parNames} element of
\code{model}.}

\item{\dots}{Not used yet.}
}
\value{
A matrix with \code{length(date)} rows and \code{3} colums.
The columns contain the location, the scale and the shape GEV parameters
in that order.
}
\description{
Compute the GEV parameters for the marginal distributions of a
\code{TVGEV} model.
}
\note{
The \code{deriv} formal argument is mainly here to maintain
a compatibility with the \code{NSGEV} method. However, the
derivatives w.r.t. the parameter vector \code{psi} are obtained
simply by using the \code{X} elements of the object \code{model}.
}
\seealso{
The \code{\link{GEV}} for the GEV probability functions.
}
