% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantMax.R
\name{qMax.TVGEV}
\alias{qMax.TVGEV}
\title{Values of the Quantile Functon of the Maximum over a
    Period of Time}
\usage{
qMax.TVGEV(
  object,
  p,
  date = NULL,
  psi = NULL,
  deriv = FALSE,
  hessian = FALSE,
  trace = 0,
  ...
)
}
\arguments{
\item{object}{A \code{TVGEV} object.}

\item{p}{A vector of probability values at which the quantile
function is to be evaluated.}

\item{date}{A vector that can be coerced to the class
\code{"Date"} which defines the period of time.}

\item{psi}{Optional vector of parameters. By default, the
estimated vector of parameters \code{coef(object)} is used.}

\item{deriv}{Logical. If \code{TRUE} the gradient of the quantile
function will be computed and returned as the
\code{"gradient"} attribute of the result.}

\item{hessian}{Logical, If \code{TRUE} the Hessian of the quantile
function will be computed and returned as the \code{"hessian"}
attribute of the result.}

\item{trace}{Integer level of verbosity.}

\item{...}{Not used yet.}
}
\value{
A vector of values for the quanitle function. Beside the
    gradient (w.r.t. the parameters) returned as the attribute
    named \code{"gradient"}, the Hessian can also be returned as
    the attribute \code{"Hessian"} of the result.
}
\description{
Compute the quantiles of the maximum over a period of time.
}
\note{
The \code{\link{quantMax}} method can be used to compute the
    quantiles and confidence intervals: delta or profile
    likelihood.
}
\section{Caution}{
 This function may be renamed in a future version.
}

\references{
See the Computing Details vignette of the package for
    the computation of the derivatives.
}
\seealso{
The \code{\link{quantMax.TVGEV}} method.
}
