% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSGEV.R
\name{quantile.NSGEV}
\alias{quantile.NSGEV}
\title{Compute \code{NSGEV} Quantiles}
\usage{
\method{quantile}{NSGEV}(x, probs = c(0.9, 0.95, 0.99), data = NULL,
  psi = NULL, ...)
}
\arguments{
\item{x}{A \code{NSGEV} object.}

\item{probs}{Vector of probabilities.}

\item{data}{A data frame with the covariates. If \code{NULL}, the
data frame affached to the model given in \code{object} is used.}

\item{psi}{Vecor of model coefficients.}

\item{...}{Not used yet.}
}
\value{
A matrix of quantiles, with one row for each observation
of \code{data} and one column by probability.
}
\description{
Compute \code{NSGEV} quantiles.
}
\examples{
df <- data.frame(t = 1:10)
## model structure
ns <- NSGEV(formulas = list("loc" = ~ alpha + beta * t,
                            "scale" = ~ delta,
                            "shape" = ~ xi),
             data = df)
psi <- c("alpha" = 1, "beta" = 0.1, "delta" = 0.6, "xi" = 0.06)
q <- quantile(ns, psi = psi) 
matplot(x = df$t, y = q, type = "l", lty = 1:3, lwd = 2,
        main = "model quantiles")
legend("bottomright", legend = colnames(q), lty = 1:3, col = 1:3,
        lwd = 2)
}
\author{
Yves Deville
}
