% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTVGEV.R
\name{quantile.TVGEV}
\alias{quantile.TVGEV}
\title{Compute Quantiles for a \code{TVGEV} Object}
\usage{
\method{quantile}{TVGEV}(x, probs = c(0.9, 0.95, 0.99), date = NULL,
  psi = NULL, ...)
}
\arguments{
\item{x}{A \code{TVGEV} object representing a time-varying GEV
model.}

\item{probs}{Vector of probabilities.}

\item{date}{An object that can be coerced to the class
\code{"Date"}. The default \code{NULL} corresponds to using
the dates attached to \code{object}.}

\item{psi}{An optional vector of parameters for \code{object}.}

\item{...}{Not used yet.}
}
\value{
An object with class \code{quantile.TVGEV} inheriting from
\code{"matrix"}. This is essentially a matrix with one column by
quantile, but it has a \code{"date"} attributes than can be used
for plotting.
}
\description{
Compute quantiles for a \code{TVGEV} object representing a
time-varying model.
}
\examples{
example(TVGEV)
q <- quantile(res1)
autoplot(q)

}
