% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantMax.R
\name{quantMax}
\alias{quantMax}
\title{Quantile of a Random Maximum}
\usage{
quantMax(object, ...)
}
\arguments{
\item{object}{An fitted model such as crated by \code{\link{TVGEV}}.}

\item{...}{Arguments for methods.}
}
\value{
The quantile of a random maximum.
}
\description{
Quantile of a Random Maximum
}
