% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantMax.R
\name{quantMaxFun}
\alias{quantMaxFun}
\title{Quantile Function for a Random Maximum}
\usage{
quantMaxFun(object, ...)
}
\arguments{
\item{object}{An object that can be used to compute/predict the
distribution of the random maximum.}

\item{...}{Further arguments for methods.}
}
\value{
A function (or closure) that can be used to compute the
    quantiles for arbitrary given probabilities.
}
\description{
Quantile Function for a Random Maximum
}
