% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantMax.R
\name{quantMaxFun.TVGEV}
\alias{quantMaxFun.TVGEV}
\title{Quantile Function for the Random Maximum on a given
    Period}
\usage{
\method{quantMaxFun}{TVGEV}(object, date = NULL, psi = NULL, theta = NULL, ...)
}
\arguments{
\item{object}{An object with class \code{"TVGEV"}.}

\item{date}{An object that can be coerced to the class
\code{"Date"}. If not provided, this will be taken as the date
attached to \code{object}.}

\item{psi}{An optional vector of coefficients for
\code{object}. By default the ML estimate as returned by
\code{coef(object)} is used.}

\item{theta}{An optional matrix with three columns containing GEV
parameters. The columns are in the order \emph{location},
\emph{scale} and \emph{shape}. When this argument is used
neither \code{date} not \code{psi} can be used.}

\item{...}{Not used.}
}
\value{
A function (more precisely, a closure). This function has
    a single formal argument \code{p} representing a probability
    \eqn{p}, and it returns the corresponding quantile \eqn{q}
    such that
    \eqn{\text{Pr}\{M^\star \leq q \} = p}{Pr[MStar <= q] = p} where
    \eqn{M^\star}{MStar} is the random maximum.
}
\description{
Given an object with class \code{"TVGEV"} and a collection of time
blocks (or \emph{period}) defined by \code{date}, the distribution
of the random maximum over the period \eqn{M^\star := \max_b
Y_b}{MStar = max_b Y_b} is known. The corresponding quantile
function \eqn{q_{M^\star}(m^\star)}{q_MStar(mStar)} can be
obtained.
}
\details{
This function computes the quantile by using the distribution
function \eqn{F_{M^\star}}{F_M} returned by
\code{\link{cdfMaxFun.TVGEV}} and the \code{\link[stats]{uniroot}}
function to solve the equation \eqn{F_{M^\star}(q) = p}{F_M(q) =
p}. By contrast \code{quantMax.TVGEV} computes the values of the
distribution values \eqn{p_i} corresponding to a grid of quantiles
\eqn{q_i} and interpolates to find the quantiles corresponding to
the probability values provided by the user. So small differences
will exist in the results.
}
\section{Caution}{
 When \code{theta} is given \code{model} is not
    used. The quantile function is simply that of the maximum of
    independent r.vs following GEV distributions with their
    parameters given by (the rows of) \code{theta}.
}

\examples{
df <- within(TXMax_Dijon, Date <- as.Date(sprintf("\%4d-01-01", Year)))
## fit a TVGEV model. Only the location parameter is TV.
res1 <- TVGEV(data = df, response = "TXMax", date = "Date",
              design = breaksX(date = Date, breaks = "1970-01-01", degree = 1),
              loc = ~ t1 + t1_1970)
qf <- quantMaxFun(res1)
qf(c(0.90, 0.99, 0.999))

## a 'new period'
date <- as.Date(sprintf("\%4d-01-01", 2025:2054))
qfNew <- quantMaxFun(res1, date = date)
qfNew(c(0.90, 0.99, 0.999))

}
\seealso{
\code{\link{cdfMaxFun.TVGEV}} for the corresponding
    quantile function (or closure).
}
