% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residTVGEV.R
\name{residuals.TVGEV}
\alias{residuals.TVGEV}
\alias{resid.TVGEV}
\title{Generalised Residuals for a \code{TVGEV} Model}
\usage{
\method{residuals}{TVGEV}(object, type = c("gumbel", "exp", "unif", "gev4"), ...)
}
\arguments{
\item{object}{A \code{TVGEV} object.}

\item{type}{The approximate distribution wanted. The choices
\code{c("gumbel", "exp", "unif", "gev4")} correspond to the
following distributions: standard Gumbel, the standard
exponential, the standard uniform, and the GEV with shape
\eqn{-0.25} distributions. Partial matching is allowed.}

\item{...}{Not used yet.}
}
\value{
A vector of generalised residuals which should
    \emph{approximately} be independent and \emph{approximately}
    follow the target distribution: standard Gumbel, standard
    exponential or standard uniform, ... depending on the value of
    \code{type}.
}
\description{
Generalised Residuals for a \code{TVGEV} model.
}
\details{
The generalised residuals are obtained by applying to each
observation \eqn{y_b} an increasing function depending on the
(estimated) GEV parameters \eqn{\boldsymbol{\theta}_b}{\theta_b}.
}
\note{
The upper 95\% quantile of the standard Gumbel and the
    standard exponential is close to \eqn{3} which can be used to
    gauge "large residuals". Using \code{type = "gumbel"} seems
    better to diagnose abnormally small residuals as may result
    from abnormally small block maxima. The generalised residuals
    have no physical dimension.
}
\examples{
df <- within(TXMax_Dijon, Date <- as.Date(sprintf("\%4d-01-01", Year)))
tv <- TVGEV(data = df, response = "TXMax", date = "Date",
            design = breaksX(date = Date, breaks = "1970-01-01", degree = 1),
            loc = ~ t1 + t1_1970)
e <- resid(tv)
plot(e)
## ggplot alternative
autoplot(e)
## plot the residual against the fitted location. Use 'as.numeric'
## on the residuals to build a similar ggplot
mu <- tv$theta[ , "loc"]
plot(mu, e, type = "p", pch = 16, col = "darkcyan",
     main = "generalised residuals against 'loc'")
}
\references{
Cox, D.R. and Snell E.J. (1968) "A General Definition
of Residuals".  \emph{JRSS Ser. B}, \bold{30}(2), pp. 248-275,
\doi{10.1111/j.2517-6161.1968.tb00724.x}.

Katz, R.W. and Parlange, M. B. and Naveau, P. (2002)
"Statistics of extremes in hydrology", \emph{Advances in Water Resources},
\bold{25}(8), pp. 1287-1304, \doi{10.1016/S0309-1708(02)00056-8}.

Panagoulia, D. and Economou, P. and Caroni, C. (2014)
"Stationary and Nonstationary Generalized Extreme Value Modelling of
 Extreme Precipitation over a Mountainous Area under Climate Change".
\emph{Environmetrics} \bold{25}(1), pp. 29-43, \doi{10.1002/env.2252}.
}
