% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{rNames}
\alias{rNames}
\title{Generates Random Names}
\usage{
rNames(n, nchar = 3L)
}
\arguments{
\item{n}{Number of names.}

\item{nchar}{Maximal number of characters for each name.}
}
\value{
A character vector of valid names.
}
\description{
Generates random names.
}
\details{
This function is a utility function for tests concerning the
parsing of formulas.
}
\examples{
set.seed(31415)
rNames(4)
rNames(4, nchar = 6)
}
