% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSGEV.R
\name{simulate.NSGEV}
\alias{simulate.NSGEV}
\title{Simulate Paths from a \code{NSGEV} Model Conditional on the
Covariates}
\usage{
\method{simulate}{NSGEV}(object, nsim = 1, seed = NULL, data = NULL, psi = NULL, ...)
}
\arguments{
\item{object}{A NSGEV model.}

\item{nsim}{Number of simulations (paths).}

\item{seed}{Not used yet.}

\item{data}{The data frame of covariates. By default, the model
data frame is used.}

\item{psi}{Vector of model parameters. By default, the model
parameters are used.}

\item{...}{Not used yet.}
}
\value{
A matrix with \code{nrow(data)} rows and \code{nsim}
columns, each column representing a time.
}
\description{
Simulate paths from a NSGEV model conditional on the
covariates.
}
\examples{
df <- data.frame(t = 1:10)
## model structure
ns <- NSGEV(formulas = list("loc" = ~ alpha + beta * t,
                            "scale" = ~ delta,
                            "shape" = ~ xi),
             data = df)
df.new <- data.frame(t = 11:20)
psi <- c("alpha" = 1, "beta" = 0.1, "delta" = 0.6, "xi" = 0.06)
ysim <- simulate(ns, nsim = 20, psi = psi)
ysim.new <- simulate(ns, nsim = 20, data = df.new, psi = psi)
matplot(df$t, ysim, type = "l", xlab = "t",
        col = "orange", xlim = c(1, 20))
matlines(df.new$t, ysim.new, type = "l", col = "firebrick")
}
\seealso{
\code{\link[nieve]{rGEV}} to simulate from varying GEV
parameters.
}
\author{
Yves Deville
}
