% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEV.R
\name{simulate.TVGEV}
\alias{simulate.TVGEV}
\title{Simulate from a \code{TVGEV} Object}
\usage{
\method{simulate}{TVGEV}(object, nsim = 1, seed = NULL, newdate = NULL, psi = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"TVGEV"} representing a
Time-Varying GEV model.}

\item{nsim}{Number of simulated}

\item{seed}{Not used yet.}

\item{newdate}{A vector with class \code{"Date"} or that can be
coerced to this class. The default \code{NULL} leads to using the
date used in \code{object}.}

\item{psi}{A vector}

\item{...}{Not used yet.}
}
\value{
A matrix with \code{nsim} columns and one row by date.
This matrix is given a special S3 class \code{"simulate.TVGEV"},
mainly in order to facilitate plotting.
}
\description{
Random simulation from a \code{TVGEV} object.
}
\examples{
example(TVGEV)
sim <- simulate(res2, nsim = 200)
plot(sim)

}
