% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bts.R
\name{[.bts}
\alias{[.bts}
\title{Extract Parts of a \code{bts} Object}
\usage{
\method{[}{bts}(x, ..., drop = FALSE)
}
\arguments{
\item{x}{A \code{bts} object.}

\item{...}{Elements to extract or replace.}

\item{drop}{\code{Logical}. If \code{TRUE} the result is coerced to the
lowest possible dimension.}
}
\description{
Extracts parts of a \code{bts} object.
}
