% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{summary.NSGEV}
\alias{summary.NSGEV}
\title{Summary method for NSGEV objects}
\usage{
\method{summary}{NSGEV}(object, ...)
}
\arguments{
\item{object}{A \code{NSGEV} object.}

\item{...}{Not used yet.}
}
\value{
A list with the elements of \code{objects}
and some more that can be displayed when \code{summary}
is invoked.
}
\description{
Summary method for \code{NSGEV} objects.
}
