% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEV.R
\name{summary.TVGEV}
\alias{summary.TVGEV}
\title{\code{summary} Method for the Class \code{TVGEV}.}
\usage{
\method{summary}{TVGEV}(object, ...)
}
\arguments{
\item{object}{An object with class \code{"TVGEV"}.}

\item{...}{Not used.}
}
\value{
A list of summary statistics.
}
\description{
\code{summary} method for the class \code{TVGEV}.
}
