% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{testNumDeriv}
\alias{testNumDeriv}
\title{Test that a Provided Graident, Jacobian or Hessian is in
    Accordance with a Numeric Hessian}
\usage{
testNumDeriv(der, derNum, PREC = 0.05, type = c("gradient", "hessian"))
}
\arguments{
\item{der}{The provided vector, matrix or array of derivatives to
be checked.}

\item{derNum}{The numeric vector, matrix or array of derivatives.}

\item{PREC}{A precision parameter. A value of \code{0.05} seems
good in practice.}
}
\value{
A logical which is \code{TRUE} if the two objects are
    close enough.
}
\description{
Test that a provided object containing derivatives is in good
accordance with another similar object containing numeric
derivatives.
}
\details{
One could think of requiring a small absolute difference or a
small relative difference between the results. However, in
practice it seems that the absolute difference can be large even
with correct derivatives because these contain large values. Also
the relative difference can be large if the derivatives contain
values that are small in absolute value. However, if one of these
two condition is fulfilled, the result must be correct because
none of the condition can be true by chance.
}
\keyword{internal}
