% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{transFormula}
\alias{transFormula}
\title{Translate a formula into a NSGEV formula}
\usage{
transFormula(formula, parnm)
}
\arguments{
\item{formula}{The formula to be tranlated.}

\item{parnm}{The name of the parameter to use, i.e.
"mu,""sigma" or "xi".}
}
\value{
A formula with the parameters displayed, suitably
named. The parameter names are attached to the result as an
attribute.
}
\description{
Translate a standard formula into a NSGEV formula.
}
\examples{

transFormula(~time, parnm = "mu") 

}
\author{
Yves Deville
}

