% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEV.R
\name{vcov.TVGEV}
\alias{vcov.TVGEV}
\title{Variance-Covariance Matrix for a \code{TVGEV} Object.}
\usage{
\method{vcov}{TVGEV}(object, ...)
}
\arguments{
\item{object}{An object with class \code{"TVGEV"} representing a
time-varying GEV model.}

\item{...}{Not used.}
}
\value{
The variance-covariance matrix.
}
\description{
Variance-covariance matrix for a \code{TVGEV} object.
}
