% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bts.R
\name{window.bts}
\alias{window.bts}
\title{Extracts a Subset of a \code{bts} Object Observed
Between the Dates \code{start} and \code{end}}
\usage{
\method{window}{bts}(x, start, end, extend = FALSE, ...)
}
\arguments{
\item{x}{A \code{bts} object.}

\item{start, end}{Objects coerced into a \code{Date} giving the start
and end.}

\item{extend}{Logical.  If \code{TRUE}, the ‘start’ and ‘end’
values are allowed to extend the series.  If \code{TRUE}, attempts
to extend the series give an error.}

\item{...}{Not used}
}
\value{
A \code{bts} object.
}
\description{
Extracts a subset of a \code{bts} object \code{x} observed between the
dates \code{start} and \code{end}.
}
