% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{coef.NSGEV}
\alias{coef.NSGEV}
\title{Coefficients of a \code{NSGEV} object}
\usage{
\method{coef}{NSGEV}(object, type = c("psi", "theta"), ...)
}
\arguments{
\item{object}{A \code{NSGEV} object.}

\item{type}{Character. When \code{"psi"}, the vector of model
parameters is returned. When instead \code{type} is \code{"theta"},
the matrix of GEV parameters is returned, with one row by block
(or observation) and one column for each of the GEV parameters
\code{"loc"}, \code{"scale"} and \code{"shape"}.}

\item{...}{Not used yet.}
}
\value{
Vector \eqn{\mathbf{\psi}}{\psi} of coefficients, or
matrix with the GEV parameters \eqn{\mathbf{\theta}_i}{\theta_i}
as its rows.
}
\description{
Extract the vector of coefficients from a NSGEV object.
}
