% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEV.R
\name{MLE.TVGEV}
\alias{MLE.TVGEV}
\title{Maximum Likelihood Estimation of a \code{TVGEV} Model}
\usage{
MLE.TVGEV(object, y = NULL, psi0 = NULL, estim = c("optim", "nloptr"),
  parTrack = FALSE)
}
\arguments{
\item{object}{A (possibly incomplete) \code{TVGEV} object.}

\item{y}{A numeric vector giving the response to be used.}

\item{psi0}{Numeric vector of initial values for the parameters.}

\item{estim}{Character giving the optimisation to be used.}

\item{parTrack}{\code{Logical}. If \code{TRUE}, all parameters at
which an evaluation of the log-Likelihood will be stored and
returned.}
}
\value{
A list with elements that can be copied into those
of a \code{TVGEV} object.
}
\description{
Maximum Likelihood Estimation of a \code{TVGEV} model.
}
\section{Caution}{
 For now it is assumed that the shape parameter
is constant. This assumtion is used to set the bounds on the shape
parameter: the lower bound is \code{-0.9} and the upper bound is
\code{2.0}.
}

\author{
Yves Deville
}
