\name{NSGEV-package}
\alias{NSGEV-package}
\docType{package}
\title{

  \packageTitle{NSGEV}

}
\description{

  \packageDescription{NSGEV}

}
\details{

  The DESCRIPTION file:
  \packageDESCRIPTION{NSGEV}
  \packageIndices{NSGEV}
  
  This package is dedicated to some Non-Stationary Extreme Value models
  with Generalised Extreme Value (GEV) marginal distributions. As a
  major difference with comparable functions found in some other EV
  packages, the model parameters can be used in several of the GEV
  parameters \code{loc}, \code{scale} and \code{shape}.  A typical
  example is when a scale proportional to the location needs to be
  specified.

  A special case is Time Varying (TV) models for which specialised
  methods make sense to plot, predict and more. The class with name
  \code{"TVGEV"} is dedicated to this kind of models.

}
\author{

   \packageAuthor{NSGEV}

   Maintainer: \packageMaintainer{NSGEV}
}
\references{
  
  Coles, S. (2001) \emph{Introdution to Statistical Modelling of Extreme
  Values}, Springer.

}
\keyword{package}
%% \seealso{}
% \examples{}
