% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bts.R
\name{plot.bts}
\alias{plot.bts}
\alias{lines.bts}
\title{Time-Plot of \code{bts} Objects for Block Time Series}
\usage{
\method{plot}{bts}(x, y, gg = TRUE, col1 = "darkgray", facets = FALSE, ...)

\method{lines}{bts}(x, y, col1 = "gray", alpha = NULL, ...)
}
\arguments{
\item{x}{A \code{bts} object representing a Block Time Series,
such as the GEV parameters of a \code{TVGEV} model.}

\item{y}{Not used yet.}

\item{gg}{Logical. If \code{TRUE}, a \code{ggplot} plot will be
drawn and returned by the function. If \code{FALSE} a standard
plot is shown and the function returns nothing.}

\item{col1}{A colour which is used in the case where \code{x} has
more than 12 columns. In this case, all the curves are drawn using
the colour \code{col1}.}

\item{facets}{Logical. If \code{True}, the columns of \code{x} will
be shown each in a facet. Only works if the number of columns is
between 2 and 12.}

\item{...}{Further parameter passed to graphical functions.}

\item{alpha}{Level of opacity. By default, the opacity is lowered
when the number of time series is increased.}
}
\value{
An object inheriting from \code{"ggplot"} if \code{gg}
is \code{TRUE}.
}
\description{
Time-plot of a \code{bts} object.
}
\note{
When a plot is built with \code{gg = TRUE}, using the
\code{lines} method later will have no effect, because this method
is designed to work on standard graphics, and not with graphics
produced by \pkg{ggplot2}.
}
\examples{
example(TVGEV)
plot(coef(res1, type = "theta"))
}
\seealso{
\code{\link{quantile.TVGEV}}, \code{\link{mean.TVGEV}} for
some examples of \code{"bts"} objects and their use in plots.
}
