% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEV.R
\name{plot.simulate.TVGEV}
\alias{plot.simulate.TVGEV}
\title{Plot Paths Simultated from a \code{TVGEV} object.}
\usage{
\method{plot}{simulate.TVGEV}(x, y, col = "gray", alpha = NULL, ...)
}
\arguments{
\item{x}{A \code{TVGEV} object}

\item{y}{Not used.}

\item{col}{Color to be used.}

\item{alpha}{Opacity level.}

\item{...}{Other arguments to be passed to \code{plot}.}
}
\value{
Nothing.
}
\description{
Plot Paths Simultated from a \code{TVGEV} object
}
\seealso{
\code{\link{simulate.TVGEV}}.
}
