% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictUncond.R
\name{predictUncond}
\alias{predictUncond}
\title{Prediction of Unconditional or Non-Stationary Return Levels
for a \code{TVGEV} or a \code{NSGEV} Object}
\usage{
predictUncond(
  object,
  period = NULL,
  level = 0.95,
  newdata = NULL,
  newdate = NULL,
  newdateFrom = NULL,
  resample = FALSE,
  RLType = c("exceed", "average"),
  confintMethod = c("delta", "none", "boot", "proflik"),
  out = c("data.frame", "array"),
  trace = 0,
  ...
)
}
\arguments{
\item{object}{A \code{TVGEV} or a \code{NSGEV} model object.}

\item{period}{A numeric vector giving the periods at which the
Return Levels will be computed.}

\item{level}{The Confidence Levels at which the CI will be
computed.  For \code{"proflik"} CI, only one level can be given
for now.}

\item{newdata}{Only for a \code{NSGEV} model. A data frame
containing the covariates needed. Each row represents a block with
unit duration.}

\item{newdate}{Only for a \code{TVGEV} model. This argument is
essentially for compatibility with the \code{predict} method for
unconditional prediction. A vector that can be coerced to the
\code{"Date"} class. This vector define \eqn{B} blocks used to
compute the return levels. The standard use is is with consecutive
blocks, so the specification should \emph{use the}
\code{newdateFrom} \emph{formal argument rather than} \code{newdate}.}

\item{newdateFrom}{Only for a \code{TVGEV} model. A vector with
length one that can be coerced to the \code{"Date"} class. This
gives the first block of a period of \eqn{B} successive blocks
used to compute the Return Levels, where \eqn{B} is the maximal
period defined in \code{period}.}

\item{resample}{Not used yet. See the \code{\link{RL}}
function. For now, the RL period that are greater than
\code{nrow(newdata)} are discarded in the computation.}

\item{RLType}{The type of Return Level as in \code{\link{RL}}.}

\item{confintMethod}{The type of Confidence Interval (CI) to
compute.  The value \code{"delta"} leads to using the \emph{delta
method}, and \code{"proflik"} leads to using the
profile-likelihood method.}

\item{out}{Type of output.}

\item{trace}{Level of verbosity.}

\item{...}{Not used yet.}
}
\value{
A data frame with the predicted RLs and the related CIs,
with one row by block.

This data frame can have several attributes

\itemize{

\item \code{title} is a string that can be used as title in plots.

\item \code{diagno}. When \code{confInt} is equal to
\code{"proflik"}, this element is an array providing information
about the optimisation.

\item \code{PsiStar}. When \code{confInt} is equal to
\code{"proflik"}, this element is a matrix with its row \eqn{i}
giving the value \eqn{\boldsymbol{\psi}^\star}{\psi*} of the vector
of parameters that maximizes the Return Level \eqn{\rho(T)} with
period \eqn{T = T_i} under the constraint on the log-likelihood.

}
}
\description{
Prediction of Unconditional or Non-Stationary Return Levels for a
\code{TVGEV} or a \code{NSGEV} object.
}
\details{
For a Non-Stationary GEV model, the Return Level corresponding to
a given period (e.g. 100 years) depends on the value of the
covariates. However we can make an assumption on the the
distribution of the covariates to compute an uncondtitional Return
Level as the expectation of the RL. Similarily, for Time-Varying
\code{TVGEV} models we can define the Return Level as the level
for which the expected number of exceedances on a reference period
of time is equal to one, see \code{\link{RL}}. In both cases, the RL
no longer depends of the covariates or on a specific block.
}
\note{
For the profile-likelihood method, the determination of the
confidence intervals is quite slow because a constrained
optimization problem is solved for each period.

Since most often the values of \eqn{\boldsymbol{\psi}^\star}{\psi*}
corresponding to different periods are close enough, a significant
reduction of the computing time could be achieved in the near
future by using the same value of \eqn{\boldsymbol{\psi}}{\psi}
for all the Return Periods.

Future versions might allow different durations across blocks by
using dedicated arguments. For now, it must be kept in mind that
the periods are understood as multiple of a constant block
duration. So if \code{newdata} has \code{100} rows the maximal
Return Period that can be used without resampling is \code{100}.
}
\section{Caution}{

\itemize{

\item When \code{period} contains some large values, say 100 or
more, a massive extrapolation from the model is required to
compute the corresponding RLs. This is likely to be meaningless
from a statistical point of view, because a massive extrapolation
of regression-like model does not make sense. The interest should
rather focus on a moderately large period.

\item For each period \eqn{T}, the Return Level is computed using
the \eqn{T} blocks \eqn{t_0}, \eqn{t_0 + 1}, \dots, \eqn{t_0 + T -1}
so the different periods \eqn{T} appearing on a Return Level plot
correspond to different prediction horizons. As a result, the
quantile or confidence bounds do not depend as smoothly of \eqn{T}
as they do for conditional predictions.

}
}

\examples{
example(TVGEV)
pu <- predictUncond(res2, newdateFrom = "2020-01-01")
pu
}
\author{
Yves Deville
}
