% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSGEV.R
\name{psi2theta}
\alias{psi2theta}
\title{Compute the matrix of GEV parameters from the vector of
model parameters 'psi' and the data}
\usage{
psi2theta(psi, model, data = NULL, deriv = TRUE, checkNames = TRUE)
}
\arguments{
\item{psi}{The vector of model parameters.}

\item{model}{The GEV model: an object with class \code{"NSGEV"}.}

\item{data}{The data frame of covariates. By default, the data
frame attached to the model is used.}

\item{deriv}{Logical. If true, the jacobian is computed and
returned as an attribute of the results. It is an array with
dimension \code{c(n, p, 3L)}}

\item{checkNames}{Logical. If \code{TRUE}, the names of the vector
\code{psi} are checked against the parnames of the model.}
}
\value{
A vector of parameters for the \code{NSGEV} model.
}
\description{
Compute the matrix of GEV parameters from the vector of model
parameters 'psi' and the data.
}
\examples{
df <- data.frame(t = 1:10)
fit <- NSGEV(formulas = list("loc" = ~ alpha + beta * t, "scale" = ~ delta, "shape" = ~ xi),
data = df)
df.new <- data.frame(t = 11:20)
psi <- c("alpha" = 1, "beta" = 0.01, "delta" = 0.6, "xi" = 0.06)
theta.new <- psi2theta(psi, model = fit, data = df.new)
matplot(df.new$t, theta.new, type = "b")
}
\author{
Yves Deville
}

