% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residTVGEV.R
\name{residuals.TVGEV}
\alias{residuals.TVGEV}
\alias{resid.TVGEV}
\title{Generalised Residuals for a \code{TVGEV} Model}
\usage{
\method{residuals}{TVGEV}(object, type = c("exp", "unif"), ...)
}
\arguments{
\item{object}{A \code{TVGEV} object.}

\item{type}{The approximate distribution wanted.}

\item{...}{Not used yet.}
}
\value{
A vector of generalised residuals which should
\emph{approximately} be independent and \emph{approximately}
follow the standard exponential or the uniform distribution,
depending on the value of \code{type}.
}
\description{
Generalised Residuals for a \code{TVGEV} model.
}
\note{
The upper 95\% quantile of the standard exponential is close
to \eqn{3} which can be used to gauge "large residuals".
}
\examples{
df <- within(TXMax_Dijon, Date <- as.Date(sprintf("\%4d-01-01", Year)))
tv <- TVGEV(data = df, response = "TXMax", date = "Date",
            design = breaksX(date = Date, breaks = "1970-01-01", degree = 1),
            loc = ~ t1 + t1_1970)
e <- resid(tv)
plot(e)
mu <- tv$theta[ , "loc"]
plot(mu, e, type = "p", pch = 16, col = "darkcyan",
     main = "generalised residuals against 'loc'")
}
\references{
Cox, D.R. and Snell E.J. (1968) "A General Definition
of Residuals".  \emph{JRSS Ser. B}, \bold{30}(2), pp. 248-275.

Panagoulia, D. and Economou, P. and Caroni, C. (2014)
"Stationary and Nonstationary Generalized Extreme Value Modelling of
 Extreme Precipitation over a Mountainous Area under Climate Change".
\emph{Environmetrics} \bold{25}(1), pp. 29-43.
}
