% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{rnames}
\alias{rnames}
\title{Generates random names}
\usage{
rnames(n, nchar = 3L)
}
\arguments{
\item{n}{Number of names.}

\item{nchar}{Maximal number of characters for each name.}
}
\value{
A character vector of valid names.
}
\description{
Generates random names.
}
\details{
This function is a utilisty function for tests concerning the
parsing of formulas.
}
\examples{
set.seed(31415)
rnames(4)
rnames(4, nchar = 6)
}

