% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{selectDate}
\alias{selectDate}
\title{Select (nearly) a Given Number of Dates}
\usage{
selectDate(date, n = 3L)
}
\arguments{
\item{date}{A vector with class\code{"Date"} or that can be coerced
to this class.}

\item{n}{Target number of selected dates.}
}
\value{
A vector of 'round' dates having nearly length \code{n}
and covering approximately the range of \code{date}.
}
\description{
Select (nearly) a given number of dates.
}
